// react-component-library/src/core/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  bigint: any;
  timestamptz: any;
  smallint: any;
  uuid: any;
};
// ...
export type SmbAttachments = {
  __typename?: 'smb_attachments';
  attachment: Scalars['String'];
  credits?: Maybe<Scalars['String']>;
  primary?: Maybe<Scalars['Boolean']>;
};
// ...
export type SmbAttributeTranslations = {
  __typename?: 'smb_attribute_translations';
  /** An object relationship */
  attribute: SmbAttributes;
  attribute_key: Scalars['String'];
  /** An object relationship */
  language: SmbLanguage;
  value?: Maybe<Scalars['String']>;
};
// ...
export type SmbAttributes = {
  __typename?: 'smb_attributes';
  datatype: Scalars['String'];
  key: Scalars['String'];
};
// ...
export type SmbHighlights = {
  __typename?: 'smb_highlights';
  /** An object relationship */
  object: SmbObjects;
  object_id: Scalars['bigint'];
};

/** aggregated selection of "smb.highlights" */
export type SmbHighlightsAggregate = {
  __typename?: 'smb_highlights_aggregate';
  aggregate?: Maybe<SmbHighlightsAggregateFields>;
  nodes: Array<SmbHighlights>;
};

/** aggregate fields of "smb.highlights" */
export type SmbHighlightsAggregateFields = {
  __typename?: 'smb_highlights_aggregate_fields';
  avg?: Maybe<SmbHighlightsAvgFields>;
  count?: Maybe<Scalars['Int']>;
  max?: Maybe<SmbHighlightsMaxFields>;
  min?: Maybe<SmbHighlightsMinFields>;
  stddev?: Maybe<SmbHighlightsStddevFields>;
  stddev_pop?: Maybe<SmbHighlightsStddevPopFields>;
  stddev_samp?: Maybe<SmbHighlightsStddevSampFields>;
  sum?: Maybe<SmbHighlightsSumFields>;
  var_pop?: Maybe<SmbHighlightsVarPopFields>;
  var_samp?: Maybe<SmbHighlightsVarSampFields>;
  variance?: Maybe<SmbHighlightsVarianceFields>;
};
// ...
export type SmbLanguage = {
  __typename?: 'smb_language';
  lang: Scalars['String'];
};
// ...
export type SmbObjects = {
  __typename?: 'smb_objects';
  /** An array relationship */
  attachments: Array<SmbAttachments>;
  /** An array relationship */
  attribute_translations: Array<SmbAttributeTranslations>;
  created_at: Scalars['timestamptz'];
  exhibition_space?: Maybe<Scalars['String']>;
  /** An array relationship */
  highlights: Array<SmbHighlights>;
  /** An aggregated array relationship */
  highlights_aggregate: SmbHighlightsAggregate;
  id: Scalars['bigint'];
  /** An array relationship */
  topics_objects: Array<SmbTopicsObjects>;
  /** An array relationship */
  tours_objects: Array<SmbToursObjects>;
  updated_at: Scalars['timestamptz'];
};
// ...
export type SmbTopics = {
  __typename?: 'smb_topics';
  has_slide: Scalars['Boolean'];
  id: Scalars['bigint'];
  /** An array relationship */
  objects: Array<SmbTopicsObjects>;
  preview_image: Scalars['String'];
  /** An array relationship */
  topics_translations: Array<SmbTopicsTranslations>;
};
// ...
export type SmbTopicsObjects = {
  __typename?: 'smb_topics_objects';
  created_at: Scalars['timestamptz'];
  id: Scalars['bigint'];
  /** An object relationship */
  object: SmbObjects;
  objects_id: Scalars['bigint'];
  /** An object relationship */
  topic: SmbTopics;
  topics_id: Scalars['bigint'];
  updated_at: Scalars['timestamptz'];
};
// ...
export type SmbTours = {
  __typename?: 'smb_tours';
  directions: Scalars['String'];
  duration: Scalars['smallint'];
  id: Scalars['bigint'];
  museum: Scalars['String'];
  number: Scalars['Int'];
  preview_image: Scalars['String'];
  /** An array relationship */
  tours_objects: Array<SmbToursObjects>;
  /** An array relationship */
  tours_translations: Array<SmbToursTranslation>;
};
// ...
export type SmbToursObjects = {
  __typename?: 'smb_tours_objects';
  /** An object relationship */
  object: SmbObjects;
  object_id: Scalars['bigint'];
  room?: Maybe<Scalars['String']>;
  sequence: Scalars['Int'];
  /** An object relationship */
  tour: SmbTours;
  tour_id: Scalars['bigint'];
  /** An array relationship */
  tours_objects_translations: Array<SmbToursObjectsTranslation>;
};
// ...
export type SmbToursObjectsTranslation = {
  __typename?: 'smb_tours_objects_translation';
  abstract: Scalars['String'];
  description: Scalars['String'];
  /** An object relationship */
  language: SmbLanguage;
};
// ...
export type SmbToursTranslation = {
  __typename?: 'smb_tours_translation';
  abstract: Scalars['String'];
  description: Scalars['String'];
  /** An object relationship */
  language: SmbLanguage;
  subtitle: Scalars['String'];
  title: Scalars['String'];
};